#783 renaming dark to dark_mode
