+++ nicegui/elements/dark_mode.js
